LayoutTests/storage/indexeddb/value-undefined-expected.txt

Test IndexedDB undefined as record value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;
PASS Boolean(indexedDB && IDBCursor && IDBDatabase && IDBDatabaseError && IDBDatabaseException && IDBFactory && IDBIndex && IDBKeyRange && IDBObjectStore && IDBRequest && IDBTransaction) is true

indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
objectStore = db.createObjectStore('foo');
result = objectStore.add(undefined, Infinity);
result = objectStore.get(Infinity);
value = event.target.result;
PASS value is undefined
result = objectStore.openCursor();
cursor = event.target.result;
PASS cursor.key is Infinity
PASS cursor.value is undefined
PASS successfullyParsed is true

TEST COMPLETE
